Integrate shared agent memory and harden terminal workspace UX
Include persistence and daemon reliability improvements, collaboration safeguards, project lifecycle controls, and UI refinements.

Add provider-neutral memory CLI/MCP tools, landing page, architecture documentation, and CI/release packaging updates. Ignore local browser captures and Cursor hook state.

+++ src/terminal/panel/tests.rs
@@ -156,10 +156,10 @@ fn zoomed_out_panel_roundings_fit_visible_header_and_body() {
 
     let roundings = panel_roundings(screen_rect, title_rect, body_rect);
 
-    assert!(roundings.panel.nw <= title_rect.height() * 0.5);
-    assert!(roundings.panel.ne <= title_rect.height() * 0.5);
-    assert!(roundings.panel.sw <= body_rect.height() * 0.5);
-    assert!(roundings.panel.se <= body_rect.height() * 0.5);
+    assert!(f32::from(roundings.panel.nw) <= title_rect.height() * 0.5);
+    assert!(f32::from(roundings.panel.ne) <= title_rect.height() * 0.5);
+    assert!(f32::from(roundings.panel.sw) <= body_rect.height() * 0.5);
+    assert!(f32::from(roundings.panel.se) <= body_rect.height() * 0.5);
     assert_eq!(roundings.title.nw, roundings.panel.nw);
     assert_eq!(roundings.body.sw, roundings.panel.sw);
 }
